[FIX] rename event
