+++ public/erc1155/446.json
@@ -1,1 +1,1 @@
-{"description":"# Description\n\nThe beckoning cat. Pull it's arm and good luck will come. A special event item from Chinese New Year!\n\n### Contributor\n\nSunflower Land is a community game built by a hundreds of developers and artists across the globe.\nCome join us on [Github](https://github.com/sunflower-land/sunflower-land)","decimals":0,"external_url":"https://docs.sunflower-land.com/getting-started/about","image_url":"https://sunflower-land.com/play/erc1155/images/446.gif","attributes":[{"trait_type":"Purpose","value":"Decoration"}],"name":"Maneki Neko"}
+{"description":"# Description\n\nThe beckoning cat. Pull it's arm and good luck will come. A special event item from Lunar New Year!\n\n### Contributor\n\nSunflower Land is a community game built by a hundreds of developers and artists across the globe.\nCome join us on [Github](https://github.com/sunflower-land/sunflower-land)","decimals":0,"external_url":"https://docs.sunflower-land.com/getting-started/about","image_url":"https://sunflower-land.com/play/erc1155/images/446.gif","attributes":[{"trait_type":"Purpose","value":"Decoration"}],"name":"Maneki Neko"}
