+++ src/assets/events/lunar-new-year/incense_left.gif
