+++ src/assets/events/lunar-new-year/lamp_npc.gif
