+++ src/assets/events/lunar-new-year/luck_sign.png
