+++ src/assets/events/lunar-new-year/new_year_island.png
