+++ src/features/game/types/quests.ts
@@ -7,8 +7,8 @@ export type QuestName =
   | "Fruit Quest 2"
   | "Fruit Quest 3"
   | "Fruit Quest 4"
-  | "Chinese New Year Quest 1"
-  | "Chinese New Year Quest 2";
+  | "Lunar New Year Quest 1"
+  | "Lunar New Year Quest 2";
 
 export type Quest = {
   description: string;
@@ -59,7 +59,7 @@ export const QUESTS: Record<QuestName, Quest> = {
     deadline: new Date(Date.now() + 10000000000).toISOString(),
   },
 
-  "Chinese New Year Quest 1": {
+  "Lunar New Year Quest 1": {
     description: "Collect 25 Red Envelopes",
     progress: (gameState: GameState) =>
       gameState.inventory["Red Envelope"]?.toNumber() || 0,
@@ -68,7 +68,7 @@ export const QUESTS: Record<QuestName, Quest> = {
     deadline: new Date("2023-01-28T00:00:00").toISOString(),
   },
 
-  "Chinese New Year Quest 2": {
+  "Lunar New Year Quest 2": {
     description: "Collect 125 Red Envelopes",
     progress: (gameState: GameState) =>
       gameState.inventory["Red Envelope"]?.toNumber() || 0,
@@ -93,6 +93,6 @@ export const BUMPKIN_QUEST_IDS: Record<QuestName | CompletedQuestName, number> =
     "Fruit Quest 2": 100006,
     "Fruit Quest 3": 100007,
     "Fruit Quest 4": 100008,
-    "Chinese New Year Quest 1": 100009,
-    "Chinese New Year Quest 2": 100010,
+    "Lunar New Year Quest 1": 100009,
+    "Lunar New Year Quest 2": 100010,
   };
